refactor(link-wizard): extract next-batch prefetch into a custom hook
Moves the queryClient.prefetchQuery logic out of LinkWizard.tsx into
usePrefetchNextBatchLinks (src/api/artistSearch/), matching the
feature-sliced API module pattern from ADR 0001.
